Componentize Mood input
Fix #10
